+++ src/Mutation.cpp
@@ -1,7 +1,7 @@
 /*
  * Authors: Gustavo V. Barroso
  * Created: 10/08/2022
- * Last modified: 05/12/2023
+ * Last modified: 30/05/2024
  *
  */
 
@@ -19,7 +19,7 @@ void Mutation::setUpMatrices_(const SumStatsLibrary& sslib)
   for(size_t i = 0; i < numPops; ++i)
   {
     size_t id = popIndices_[i];
-    std::vector<Eigen::Triplet<double>> coeffs(0);
+    std::vector<Eigen::Triplet<long double>> coeffs(0);
     coeffs.reserve(sizeOfBasis);
 
     for(auto it = std::begin(sslib.getBasis()); it != std::end(sslib.getBasis()); ++it)
@@ -32,13 +32,13 @@ void Mutation::setUpMatrices_(const SumStatsLibrary& sslib)
       if((*it)->getPrefix() == "Hl")
       {
         col = sslib.findCompressedIndex(sslib.getMoment("I")); // for a homogeneous system
-        coeffs.emplace_back(Eigen::Triplet<double>(row, col, leftFactor_ * popIdCount / 2.));
+        coeffs.emplace_back(Eigen::Triplet<long double>(row, col, leftFactor_ * popIdCount / 2.));
       }
 
       else if((*it)->getPrefix() == "Hr")
       {
         col = sslib.findCompressedIndex(sslib.getMoment("I")); // for a homogeneous system
-        coeffs.emplace_back(Eigen::Triplet<double>(row, col, popIdCount / 2.)); // ie, rightFactor_ == 1
+        coeffs.emplace_back(Eigen::Triplet<long double>(row, col, popIdCount / 2.)); // ie, rightFactor_ == 1
       }
 
       else if((*it)->getPrefix() == "pi2")
@@ -50,20 +50,20 @@ void Mutation::setUpMatrices_(const SumStatsLibrary& sslib)
         auto tempLeft = tmpPi2->getLeftHetStat();
         size_t leftMult = tempLeft->countInstances(id);
         col = tempLeft->getPosition();
-        coeffs.emplace_back(Eigen::Triplet<double>(row, col, leftMult / 2.));
+        coeffs.emplace_back(Eigen::Triplet<long double>(row, col, leftMult / 2.));
 
         // introducing 2-locus Het via mutation in left locus (when right already polymorphic)
         auto tempRight = tmpPi2->getRightHetStat();
         size_t rightMult = tempRight->countInstances(id);
         col = tempRight->getPosition();
-        coeffs.emplace_back(Eigen::Triplet<double>(row, col, rightMult / 2.));
+        coeffs.emplace_back(Eigen::Triplet<long double>(row, col, rightMult / 2.));
       }
 
       else if((*it)->getPrefix() != "I" && (*it)->getPrefix() != "DD" && (*it)->getPrefix() != "Dr" && (*it)->getPrefix() != "D")
         throw bpp::Exception("Mutation::mis-specified Moment prefix: " + (*it)->getPrefix());
     }
 
-    Eigen::SparseMatrix<double> mat(sizeOfBasis, sizeOfBasis);
+    Eigen::SparseMatrix<long double> mat(sizeOfBasis, sizeOfBasis);
     mat.setFromTriplets(std::begin(coeffs), std::end(coeffs));
     mat.makeCompressed();
     mat *= getParameterValue("u_" + bpp::TextTools::toString(id));
